+++ llmfoundry/models/mpt/modeling_mpt.py
@@ -28,6 +28,7 @@ from composer.utils import dist
 
 from llmfoundry.models.layers.attention import (is_flash_v1_installed,
                                                 is_flash_v2_installed)
+from llmfoundry.models.layers.norm import NORM_CLASS_REGISTRY
 
 if is_flash_v2_installed():
     try:  # This try...except is needed because transformers requires it despite the 'if' statement above
@@ -55,17 +56,11 @@ from transformers.models.llama.modeling_llama import \
 from transformers.models.llama.modeling_llama import \
     LlamaRotaryEmbedding as HFRotaryEmbedding
 
-from llmfoundry.models.layers.attention import (ATTN_CLASS_REGISTRY,
-                                                attn_bias_shape,
+from llmfoundry.models.layers.attention import (attn_bias_shape,
                                                 build_attn_bias, gen_slopes)
 from llmfoundry.models.layers.blocks import MPTBlock
 from llmfoundry.models.layers.custom_embedding import SharedEmbedding
-from llmfoundry.models.layers.fc import FC_CLASS_REGISTRY as FC_CLASS_REGISTRY
-from llmfoundry.models.layers.ffn import \
-    FFN_CLASS_REGISTRY as FFN_CLASS_REGISTRY
-from llmfoundry.models.layers.ffn import MPTMLP as MPTMLP
 from llmfoundry.models.layers.ffn import build_ffn as build_ffn
-from llmfoundry.models.layers.norm import NORM_CLASS_REGISTRY
 from llmfoundry.models.mpt.configuration_mpt import MPTConfig
 
 # NOTE: All utils are imported directly even if unused so that
@@ -86,6 +81,10 @@ from llmfoundry.models.utils.param_init_fns import (
     generic_param_init_fn_,  # type: ignore (see note)
     MODEL_INIT_REGISTRY,
 )
+
+from llmfoundry.models.utils.act_ckpt import (get_act_ckpt_module,
+                                              get_target_block_list,
+                                              check_mapping_blocks_overlap)
 
 try:
     from llmfoundry.models.layers.flash_attn_triton import flash_attn_func as flash_attn_func
@@ -947,85 +946,6 @@ class MPTForCausalLM(MPTPreTrainedModel):
             )
             return False
 
-        def get_act_ckpt_module(mod_name: str) -> Any:
-            if mod_name.lower() == 'mptblock':
-                mod_type = MPTBlock
-            elif mod_name in ATTN_CLASS_REGISTRY:
-                mod_type = ATTN_CLASS_REGISTRY[mod_name]
-            elif mod_name in FFN_CLASS_REGISTRY:
-                mod_type = FFN_CLASS_REGISTRY[mod_name]
-            elif mod_name in NORM_CLASS_REGISTRY:
-                mod_type = NORM_CLASS_REGISTRY[mod_name]
-            else:
-                msg = ', '.join(
-                    list(ATTN_CLASS_REGISTRY.keys()) +
-                    list(FFN_CLASS_REGISTRY.keys()) +
-                    list(NORM_CLASS_REGISTRY.keys()) + ['MPTBlock'])
-                raise ValueError(
-                    f'{mod_name} (specified in activation_checkpointing_target) is not a recognized option out of available options {msg}.'
-                )
-            return mod_type
-
-        def get_target_block_list(target_blocks: Any,
-                                  max_block_idx: int) -> list:
-
-            def parse_ele_str(ele: str, max_block_idx: int) -> list:
-                to_add = None
-                if ele.startswith('first-'):
-                    assert ele[6:].isdigit(
-                    ), f'Invalid target_blocks element {ele}'
-                    to_add = list(range(min(int(ele[6:]), max_block_idx + 1)))
-                elif ele.startswith('last-'):
-                    assert ele[5:].isdigit(
-                    ), f'Invalid target_blocks element {ele}'
-                    to_add = list(
-                        range(max(max_block_idx - int(ele[5:]) + 1, 0),
-                              max_block_idx + 1))
-                elif ele.startswith('middle-'):
-                    assert ele[7:].isdigit(
-                    ), f'Invalid target_blocks element {ele}'
-                    num = int(ele[7:])
-                    start = max(max_block_idx // 2 - num // 2, 0)
-                    end = min(start + num, max_block_idx + 1)
-                    to_add = list(range(start, end))
-                elif ele.startswith('range-'):
-                    r = ele[6:].split('-')
-                    assert len(r) == 2, f'Invalid target_blocks element {ele}'
-                    start, end = int(r[0]), int(r[1])
-                    start = max(start, 0)
-                    end = min(end, max_block_idx + 1)
-                    to_add = list(range(start, end))
-                else:
-                    raise ValueError(f'Invalid target_blocks element {ele}')
-                return to_add
-
-            candidate_block_ids = []
-            if isinstance(target_blocks, int):
-                candidate_block_ids = list(range(target_blocks))
-            elif isinstance(target_blocks, list):
-                for ele in target_blocks:
-                    if isinstance(ele, int):
-                        candidate_block_ids.append(ele)
-                    elif isinstance(ele, str):
-                        to_add = parse_ele_str(ele, max_block_idx)
-                        candidate_block_ids.extend(to_add)
-                    else:
-                        raise ValueError(
-                            f'target_blocks must be a list of integers or "first-n", "middle-m", "last-k", or "range-i-j" where n, m, k, i, j are integers, but got {target_blocks}'
-                        )
-            elif isinstance(target_blocks, str):
-                target_blocks = target_blocks.replace(' ', '')
-                for ele in target_blocks.split(','):
-                    to_add = parse_ele_str(ele, max_block_idx)
-                    candidate_block_ids.extend(to_add)
-            else:
-                raise ValueError(
-                    f'target_blocks must be either a single intege, or a list of integers, or a comma separated string made of "first-n", "last-m", "middle-k", "range-i-j", or a list of mixed integers and before-mentioned strings, but got {type(target_blocks)}'
-                )
-
-            candidate_block_ids = list(set(candidate_block_ids))
-            return candidate_block_ids
-
         act_ckpt_target = getattr(self.config,
                                   'activation_checkpointing_target', None)
         act_ckpt_mod_to_blocks = {}
@@ -1051,21 +971,6 @@ class MPTForCausalLM(MPTPreTrainedModel):
             raise ValueError(
                 f'activation_checkpointing_target must be either a single string or a list or a dict, but got {type(act_ckpt_target)}'
             )
-
-        def check_mapping_blocks_overlap(mapping: dict,
-                                         max_block_idx: int) -> None:
-            all_blocks = [None] * (max_block_idx + 1)
-            for k, v in mapping.items():
-                for vv in v:
-                    if vv < 0 or vv > max_block_idx:
-                        continue
-                    else:
-                        if all_blocks[vv] is not None:
-                            raise ValueError(
-                                f'Block {vv} is assigned to both {k} and {all_blocks[vv]}.'
-                            )
-                        else:
-                            all_blocks[vv] = k
 
         check_mapping_blocks_overlap(act_ckpt_mod_to_blocks,
                                      module.max_block_idx)
